Fix: Place the invitation reminder icon in a separate file #9634
